deps/v8/test/inspector/runtime/error-preview-sideeffects-expected.txt

Checks that error preview does not run user code

Running test: ErrorStack
{
    stackCalled : false
    stackCalledOnBase : false
    stackCalledOnProxy : false
}
